chore(integer-i64): fmt + clippy cleanup after 2c rounds 5–10
Accumulated mechanical debt from widening i32→i64 across the
interpreter.  All changes are refactoring-safe auto-fixes;
zero behavioral changes confirmed by running the full test
suite before and after the cleanup (both produce the same
set of 59 failing tests — all pre-existing 2c issues, none
introduced by this commit).

Applied via `cargo fmt --all` and `cargo clippy --fix --all-
targets -- -D warnings`, then a fresh regen of `src/fill.rs`
to pick up a tweak to `default/01_code.loft` (an
`OpLengthVector` `#rust` template still used `as i64` on a
u32 input; `i64::from(...)` is what clippy wants everywhere
else post-2c).

Categories of lint fixed:

- `unnecessary_cast` (14): `x as u32` where x is already u32,
  also a couple of `u16 as u16` and `i32 as i32`.  Residue
  of earlier u32→u64 / u16→u16 widenings.
- `useless_conversion` (4): `i64::from(get_int(...))` after
  `get_int` started returning i64 directly.
- `cast_lossless` (4): `u32 as i64` → `i64::from(u32)`.
- `if_not_else` / `unnecessary_!=` (4): pure cosmetic.

Unblocks `cargo fmt --check` and `cargo clippy -D warnings`
on both debug and release profiles, letting CI gate the
`int_migrate` branch as it approaches merge.

+++ src/database/allocation.rs
@@ -876,7 +876,7 @@ impl Stores {
                 if cur == 0 {
                     return;
                 }
-                store.delete(cur as u32);
+                store.delete(cur);
                 store.set_u32_raw(rec.rec, rec.pos, 0);
             }
             Parts::Struct(fields) | Parts::EnumValue(_, fields) => {
@@ -904,14 +904,14 @@ impl Stores {
                     self.remove_claims(
                         &DbRef {
                             store_nr: rec.store_nr,
-                            rec: cur as u32,
+                            rec: cur,
                             pos: 8 + size * i,
                         },
                         tp,
                     );
                 }
                 let store = self.store_mut(rec);
-                store.delete(cur as u32);
+                store.delete(cur);
                 store.set_u32_raw(rec.rec, rec.pos, 0);
             }
             Parts::Array(v) | Parts::Ordered(v, _) => {
